--- docs/README.md
@@ -1,1 +1,0 @@
-Location for expanded documentation.
