--- godotsteam/SCsub
@@ -1,5 +1,0 @@
-
-# SCsub
-Import('env')
-
-env.add_source_files(env.modules_sources,"*.cpp")
