--- godotsteam/register_types.h
@@ -1,3 +1,0 @@
-// Yes, the word in the middle must be the same as the module folder name
-void register_godotsteam_types();
-void unregister_godotsteam_types();
