+++ lib/steam.tscn
@@ -1,0 +1,15 @@
+[gd_scene load_steps=3 format=2]
+
+[ext_resource path="res://lib/godotsteam.tres" type="GDNativeLibrary" id=1]
+
+[sub_resource type="GDNativeScript" id=1]
+
+resource_name = "Steam"
+library = ExtResource( 1 )
+script_name = "Steam"
+
+[node name="Steam" type="Node"]
+
+script = SubResource( 1 )
+
+
